movement-service/requirements.txt
grpcio
grpcio-tools
opencv-python
mediapipe
python-dotenv